+++ src/Models/Role.php
@@ -13,6 +13,19 @@ class Role extends SpatieRole
     use SoftDeletes;
     use HasTenantScope;
 
+  /**
+     * The attributes that are mass assignable.
+     *
+     * @var array<string>
+     */
+    protected $fillable = [
+        'name',
+        'guard_name',
+        'status',
+        'label',
+        'description',
+    ];
+
     /**
      * Get the attributes that should be cast.
      *